requirements.txt
django==3.2
whitenoise==5.2.0
gunicorn==20.0.4
requests==2.4.2
django-hitcount==1.3.5